#1115 rename + refactor userInfoAuthority files
