+++ wls-briefwahl-service/src/main/java/de/muenchen/oss/wahllokalsystem/briefwahlservice/configuration/SecurityConfiguration.java
@@ -55,7 +55,7 @@ public class SecurityConfiguration {
                         .authenticated())
                 .oauth2ResourceServer(httpSecurityOAuth2ResourceServerConfigurer -> httpSecurityOAuth2ResourceServerConfigurer
                         .jwt(jwtConfigurer -> jwtConfigurer.jwtAuthenticationConverter(new JwtUserInfoAuthenticationConverter(
-                                new UserInfoAuthoritiesService(userInfoUri, restTemplateBuilder)))));
+                                new UserInfoAuthoritiesRetriever(userInfoUri, restTemplateBuilder)))));
 
         return http.build();
     }
